+++ apps/design-system/src/subjects/views/secrets/mock-account-data.json
@@ -1,0 +1,40 @@
+[
+  {
+    "uuid": "random-uuid-here",
+    "accountName": "harness-dev",
+    "companyName": "harness-dev",
+    "clusterName": "cluster1",
+    "localEncryptionEnabled": false,
+    "twoFactorAdminEnforced": false,
+    "forImport": false,
+    "cloudCostEnabled": false,
+    "salesContacts": null,
+    "licenseInfo": {
+      "accountType": "PAID",
+      "accountStatus": "ACTIVE",
+      "expireAfterDays": 0,
+      "expiryTime": 1727111595000,
+      "licenseUnits": 1003000,
+      "reasonForLicenseUpdate": null
+    },
+    "ceLicenseInfo": {
+      "licenseType": "FULL_TRIAL",
+      "expiryTime": 1656700199999
+    },
+    "licenseId": null,
+    "techStacks": [
+      {
+        "category": "Deployment Platforms",
+        "technology": "None"
+      }
+    ],
+    "subdomainUrl": null,
+    "defaultExperience": "NG",
+    "authenticationMechanism": "USER_PASSWORD",
+    "sessionTimeOutInMinutes": 1440,
+    "publicAccessEnabled": true,
+    "absoluteSessionTimeOutInMinutes": 0,
+    "status": "ACTIVE",
+    "loginAllowed": true
+  }
+]
