+++ apps/design-system/src/subjects/views/secrets/mock-org-data.json
@@ -1,0 +1,42 @@
+[
+  {
+    "organizationResponse": {
+      "organization": {
+        "identifier": "org-anon-1",
+        "name": "org-anon-1",
+        "description": "",
+        "tags": {}
+      },
+      "createdAt": 1720080149569,
+      "lastModifiedAt": 1720080149569,
+      "harnessManaged": false
+    },
+    "projectsCount": 1,
+    "connectorsCount": 1,
+    "secretsCount": 0,
+    "delegatesCount": 0,
+    "templatesCount": 0,
+    "admins": [
+      {
+        "name": "Admin User 1",
+        "email": "admin1@example.com",
+        "uuid": "uuid-admin-1",
+        "locked": false,
+        "disabled": false,
+        "externallyManaged": false,
+        "twoFactorAuthenticationEnabled": false
+      }
+    ],
+    "collaborators": [
+      {
+        "name": "Collaborator User 1",
+        "email": "collab1@example.com",
+        "uuid": "uuid-collab-1",
+        "locked": false,
+        "disabled": false,
+        "externallyManaged": false,
+        "twoFactorAuthenticationEnabled": false
+      }
+    ]
+  }
+]
