+++ apps/design-system/src/subjects/views/secrets/mock-project-data.json
@@ -1,0 +1,120 @@
+[
+  {
+    "projectResponse": {
+      "project": {
+        "orgIdentifier": "org-1",
+        "identifier": "project-1",
+        "name": "project-1",
+        "color": "#0063f7",
+        "modules": [
+          "CD",
+          "CI",
+          "CV",
+          "CF",
+          "CE",
+          "STO",
+          "CHAOS",
+          "SRM",
+          "IACM",
+          "CET",
+          "IDP",
+          "CODE",
+          "SSCA",
+          "CORE",
+          "PMS",
+          "TEMPLATESERVICE",
+          "SEI",
+          "HAR"
+        ],
+        "description": "",
+        "tags": {}
+      },
+      "createdAt": 1720080389723,
+      "lastModifiedAt": 1720080389723,
+      "isFavorite": false
+    },
+    "organization": {
+      "identifier": "org-test-1",
+      "name": "Org Test 1",
+      "description": "",
+      "tags": {}
+    },
+    "harnessManagedOrg": false,
+    "admins": [
+      {
+        "name": "Admin 1",
+        "email": "admin1@example.com",
+        "uuid": "admin1-uuid",
+        "locked": false,
+        "disabled": false,
+        "externallyManaged": false,
+        "twoFactorAuthenticationEnabled": false
+      }
+    ],
+    "collaborators": []
+  },
+  {
+    "projectResponse": {
+      "project": {
+        "orgIdentifier": "org-2",
+        "identifier": "project-2",
+        "name": "project-2",
+        "color": "#0063f7",
+        "modules": [
+          "CD",
+          "CI",
+          "CV",
+          "CF",
+          "CE",
+          "STO",
+          "CHAOS",
+          "SRM",
+          "IACM",
+          "CET",
+          "IDP",
+          "CODE",
+          "SSCA",
+          "CORE",
+          "PMS",
+          "TEMPLATESERVICE",
+          "SEI",
+          "HAR"
+        ],
+        "description": "",
+        "tags": {}
+      },
+      "createdAt": 1700827366661,
+      "lastModifiedAt": 1700827366661,
+      "isFavorite": false
+    },
+    "organization": {
+      "identifier": "org-test-2",
+      "name": "Org Test 2",
+      "description": "",
+      "tags": {}
+    },
+    "harnessManagedOrg": false,
+    "admins": [
+      {
+        "name": "Admin 2",
+        "email": "admin2@example.com",
+        "uuid": "admin2-uuid",
+        "locked": false,
+        "disabled": false,
+        "externallyManaged": false,
+        "twoFactorAuthenticationEnabled": false
+      }
+    ],
+    "collaborators": [
+      {
+        "name": "Collaborator 1",
+        "email": "collab1@example.com",
+        "uuid": "collab1-uuid",
+        "locked": false,
+        "disabled": false,
+        "externallyManaged": false,
+        "twoFactorAuthenticationEnabled": false
+      }
+    ]
+  }
+]
