+++ apps/design-system/src/subjects/views/secrets/mock-secrets-data.json
@@ -1,0 +1,128 @@
+[
+  {
+    "secret": {
+      "type": "SecretFile",
+      "name": "secret-file-1",
+      "identifier": "secret-id-1",
+      "orgIdentifier": "org-default",
+      "projectIdentifier": "project-xyz",
+      "tags": {},
+      "description": "",
+      "spec": {
+        "secretManagerIdentifier": "secret-manager",
+        "additionalMetadata": null
+      }
+    },
+    "createdAt": 1741279791821,
+    "updatedAt": 1741279791821,
+    "draft": false,
+    "governanceMetadata": null
+  },
+  {
+    "secret": {
+      "type": "SecretFile",
+      "name": "secret-file-2",
+      "identifier": "secret-id-2",
+      "orgIdentifier": "org-default",
+      "projectIdentifier": "project-xyz",
+      "tags": {},
+      "description": "",
+      "spec": {
+        "secretManagerIdentifier": "secret-manager",
+        "additionalMetadata": null
+      }
+    },
+    "createdAt": 1741213431403,
+    "updatedAt": 1741213431403,
+    "draft": false,
+    "governanceMetadata": null
+  },
+  {
+    "secret": {
+      "type": "SecretText",
+      "name": "secret-text-1",
+      "identifier": "secret-text-id-1",
+      "orgIdentifier": "org-default",
+      "projectIdentifier": "project-xyz",
+      "tags": {
+        "tag1": "",
+        "tag2": "",
+        "tag3": ""
+      },
+      "description": "desc",
+      "spec": {
+        "secretManagerIdentifier": "secret-manager",
+        "valueType": "Inline",
+        "value": null,
+        "additionalMetadata": null
+      }
+    },
+    "createdAt": 1740785306341,
+    "updatedAt": 1740785306341,
+    "draft": false,
+    "governanceMetadata": null
+  },
+  {
+    "secret": {
+      "type": "SecretText",
+      "name": "secret-text-2",
+      "identifier": "secret-text-id-2",
+      "orgIdentifier": "org-default",
+      "projectIdentifier": "project-xyz",
+      "tags": {},
+      "description": "github token",
+      "spec": {
+        "secretManagerIdentifier": "secret-manager",
+        "valueType": "Inline",
+        "value": null,
+        "additionalMetadata": null
+      }
+    },
+    "createdAt": 1719353653579,
+    "updatedAt": 1719353653579,
+    "draft": false,
+    "governanceMetadata": null
+  },
+  {
+    "secret": {
+      "type": "SecretText",
+      "name": "secret-text-3",
+      "identifier": "secret-text-id-3",
+      "orgIdentifier": "org-default",
+      "projectIdentifier": "project-xyz",
+      "tags": {},
+      "description": "",
+      "spec": {
+        "secretManagerIdentifier": "secret-manager",
+        "valueType": "Inline",
+        "value": null,
+        "additionalMetadata": null
+      }
+    },
+    "createdAt": 1710968557986,
+    "updatedAt": 1710968557986,
+    "draft": false,
+    "governanceMetadata": null
+  },
+  {
+    "secret": {
+      "type": "SecretText",
+      "name": "secret-text-4",
+      "identifier": "secret-text-id-4",
+      "orgIdentifier": "org-default",
+      "projectIdentifier": "project-xyz",
+      "tags": {},
+      "description": "",
+      "spec": {
+        "secretManagerIdentifier": "secret-manager",
+        "valueType": "Inline",
+        "value": null,
+        "additionalMetadata": null
+      }
+    },
+    "createdAt": 1700005592914,
+    "updatedAt": 1700005592914,
+    "draft": false,
+    "governanceMetadata": null
+  }
+]
