+++ packages/ui/src/views/platform/entity-reference-component.tsx
@@ -1,0 +1,134 @@
+import { useCallback } from 'react'
+
+import { Button, Icon, Input, ScrollArea, SkeletonList, StackedList } from '@/components'
+import { cn } from '@utils/cn'
+
+import { EntityReferenceList } from './entity-reference-list'
+import { BaseEntityProps, DirectionEnum, EntityRendererProps, FolderRendererProps, ScopeSelectorProps } from './types'
+
+export interface EntityReferenceProps<T extends BaseEntityProps, S = string, F = string> {
+  // Data
+  entities: T[]
+  selectedEntity: T | null
+  parentFolder: S | null
+  childFolder: F | null
+
+  // Callbacks
+  onSelectEntity: (entity: T) => void
+  onScopeChange: (direction: DirectionEnum) => void
+
+  // UI Configuration
+  showFilter?: boolean
+
+  // Custom renderers
+  renderEntity?: (props: EntityRendererProps<T>) => React.ReactNode
+  isLoading?: boolean
+}
+
+export function EntityReference<T extends BaseEntityProps, S = string, F = string>({
+  // Data
+  entities,
+  selectedEntity,
+  parentFolder,
+  childFolder,
+
+  // Callbacks
+  onSelectEntity,
+  onScopeChange,
+
+  // configs
+  showFilter = true,
+
+  // Custom renderers
+  renderEntity,
+  isLoading = false
+}: EntityReferenceProps<T, S, F>): JSX.Element {
+  const handleSelectEntity = useCallback(
+    (entity: T) => {
+      onSelectEntity?.(entity)
+    },
+    [onSelectEntity]
+  )
+
+  const handleScopeChange = useCallback(
+    (direction: DirectionEnum) => {
+      onScopeChange?.(direction)
+    },
+    [onScopeChange]
+  )
+
+  const defaultEntityRenderer = ({ entity, isSelected, onSelect }: EntityRendererProps<T>) => {
+    return (
+      <StackedList.Item
+        onClick={() => onSelect?.(entity)}
+        className={cn({ 'bg-background-4': isSelected })}
+        thumbnail={<Icon name="file" size={16} className="text-foreground-5" />}
+        actions={
+          <Button
+            variant="default"
+            size="sm"
+            onClick={() => {
+              onSelect?.(entity)
+            }}
+          >
+            Select
+          </Button>
+        }
+      >
+        <StackedList.Field title={entity.name} />
+      </StackedList.Item>
+    )
+  }
+
+  const parentFolderRenderer = ({ parentFolder, onSelect }: ScopeSelectorProps<S>) => {
+    return (
+      <StackedList.Item
+        onClick={() => onSelect?.(parentFolder)}
+        thumbnail={<Icon name="circle-arrow-top" size={16} className="text-foreground-5" />}
+      >
+        <StackedList.Field title={<span className="capitalize">{String(parentFolder)}</span>} />
+      </StackedList.Item>
+    )
+  }
+
+  const childFolderRenderer = ({ folder, onSelect }: FolderRendererProps<F>) => {
+    return (
+      <StackedList.Item
+        onClick={() => onSelect?.(folder)}
+        thumbnail={<Icon name="folder" size={16} className="text-foreground-5" />}
+      >
+        <StackedList.Field title={<span className="capitalize">{String(folder)}</span>} />
+      </StackedList.Item>
+    )
+  }
+
+  return (
+    <>
+      {isLoading ? (
+        <SkeletonList />
+      ) : (
+        <>
+          {showFilter && <Input type="text" placeholder="Search" className="mb-4" />}
+          <ScrollArea className="h-[62%]">
+            <div className="flex-1">
+              <EntityReferenceList
+                entities={entities}
+                selectedEntity={selectedEntity}
+                parentFolder={parentFolder}
+                childFolder={childFolder}
+                handleSelectEntity={handleSelectEntity}
+                handleScopeChange={handleScopeChange}
+                renderEntity={renderEntity}
+                defaultEntityRenderer={defaultEntityRenderer}
+                parentFolderRenderer={parentFolderRenderer}
+                childFolderRenderer={childFolderRenderer}
+              />
+            </div>
+          </ScrollArea>
+        </>
+      )}
+    </>
+  )
+}
+
+export default EntityReference
