+++ packages/ui/src/views/platform/index.ts
@@ -1,1 +1,2 @@
 export * from './entity-reference-component'
+export * from './types'
